Revert "공지추가,수정화면 CSS수정&JS추가,환자VO 몸무게 데이터타입(float)변경&보호자 탈퇴여부 데이터타입(boolean)변경"
This reverts commit 1502e2f096f076442917d15eca57a29fedf1cb00.

+++ AnimalHospitalProject/src/main/java/kr/human/anihospital/vo/AnimalVO.java
@@ -16,7 +16,7 @@ public class AnimalVO {
 	private String animalName;
 	private String animalType;
 	private String animalSize;
-	private float animalWeight;
+	private int animalWeight;
 	private Date animalAge;
 	private boolean animalGender;
 	private String animalImportantSymptom;
